+++ apps/block_scout_web/lib/block_scout_web/notifier.ex
@@ -3,6 +3,8 @@ defmodule BlockScoutWeb.Notifier do
   Responds to events by sending appropriate channel updates to front-end.
   """
 
+  require Logger
+
   alias Absinthe.Subscription
 
   alias BlockScoutWeb.API.V2, as: API_V2
@@ -48,6 +50,8 @@ defmodule BlockScoutWeb.Notifier do
   def handle_event(
         {:chain_event, :contract_verification_result, :on_demand, {address_hash, contract_verification_result}}
       ) do
+    log_broadcast_verification_results_for_address(address_hash)
+
     Endpoint.broadcast(
       "addresses:#{address_hash}",
       "verification_result",
@@ -60,6 +64,7 @@ defmodule BlockScoutWeb.Notifier do
   def handle_event(
         {:chain_event, :contract_verification_result, :on_demand, {address_hash, contract_verification_result, conn}}
       ) do
+    log_broadcast_verification_results_for_address(address_hash)
     %{view: view, compiler: compiler} = select_contract_type_and_form_view(conn.params)
 
     contract_verification_result =
@@ -218,6 +223,7 @@ defmodule BlockScoutWeb.Notifier do
   end
 
   def handle_event({:chain_event, :smart_contract_was_verified, :on_demand, [address_hash]}) do
+    log_broadcast_smart_contract_was_verified(address_hash)
     Endpoint.broadcast("addresses:#{to_string(address_hash)}", "smart_contract_was_verified", %{})
   end
 
@@ -227,7 +233,10 @@ defmodule BlockScoutWeb.Notifier do
     })
   end
 
-  def handle_event(_), do: nil
+  def handle_event(event) do
+    Logger.warning("Unknown broadcasted event #{inspect(event)}.")
+    nil
+  end
 
   def fetch_compiler_version(compiler) do
     case CompilerVersion.fetch_versions(compiler) do
@@ -480,5 +489,13 @@ defmodule BlockScoutWeb.Notifier do
     for {address_hash, elements} <- grouped do
       Endpoint.broadcast("addresses:#{address_hash}", event, %{map_key => elements})
     end
+  end
+
+  defp log_broadcast_verification_results_for_address(address_hash) do
+    Logger.info("Broadcast smart-contract #{address_hash} verification results")
+  end
+
+  defp log_broadcast_smart_contract_was_verified(address_hash) do
+    Logger.info("Broadcast smart-contract #{address_hash} was verified")
   end
 end