+++ apps/explorer/lib/explorer/smart_contract/vyper/publisher_worker.ex
@@ -3,6 +3,8 @@ defmodule Explorer.SmartContract.Vyper.PublisherWorker do
   Background smart contract verification worker.
   """
 
+  require Logger
+
   use Que.Worker, concurrency: 5
 
   alias Explorer.Chain.Events.Publisher, as: EventsPublisher
@@ -33,6 +35,8 @@ defmodule Explorer.SmartContract.Vyper.PublisherWorker do
         {:error, changeset} ->
           {:error, changeset}
       end
+
+    Logger.info("Smart-contract #{address_hash} verification: broadcast verification results")
 
     if conn do
       EventsPublisher.broadcast([{:contract_verification_result, {address_hash, result, conn}}], :on_demand)